pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hpd-agent"
version = "0.1.0"
description = "HPD-Agent CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain-openai>=0.1.0",
  "langgraph>=0.2.0",
  "langgraph-checkpoint>=2.0.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "tiktoken>=0.7.0",
  "dacite>=0.8.0",
  "prompt_toolkit>=3.0.0",
  "rich>=13.0.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
hpd = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
